Standards-based CSS sanitizer
This adds the classes needed to actually validate a stylesheet against
the grammar of many CSS3 modules, properly stripping out unrecognized
properties, rules, and so on.

The choices of which modules to include is somewhat arbitrary, I mostly
just grabbed anything I saw that seemed to have a decent level of
browser support.

Change-Id: I101b6bceed95da3ba8debcb4ccb59a832fc6c9ac