Remove "preprocessor comments"
We had thought we'd need them, but it turns out to make more sense to
handle them before sanitizing rather than after.

Since keeping them leaves a lot of open questions as to where exactly
they're allowed to appear and means all other code has to be sure to
skip them when processing a token stream, let's remove them for now.
They can be brought back if we find a case where we need them.

Change-Id: Id26211fcb5197eaad617b73409a5388d6de102a2