Parser: Add a ComponentValue nesting depth limit
It's really easy to make the parser hit a stack depth limit, all you
need to do is throw it a file with a few thousand open-braces that are
interpreted as nested SimpleBlocks. You can do the same with
CSSFunctions, or a mix of the two. So, we introduce a recursion limit,
arbitrarily set at 100 since that seems like it should be enough for any
sane CSS stylesheet.

Note that, at the parser level, the only nested constructs are
ComponentValues. Rules and such are just the top-level rule with a block
containing a not-yet-differentiated list of ComponentValues whose
interpretation later will depend on the context.

Change-Id: I9a8a239b43aa552306b24fa25a2725d1db23e54f