Add hoisting option to StyleRuleSanitizer selector handling
This is technically a breaking change in that StyleRuleSanitizer's
prependSelectors option does not need to end with space (and using
a prefix that's not separated by whitespace from the prefixed
part is not possible anymore). The extra whitespace generated
with an old-style prependSelectors value is harmless though, and
there was never any use case for not separating prepended
selectors with a descendant combinator.

Bug: T197617
Co-Authored-By: Brad Jorsch <bjorsch@wikimedia.org>
Change-Id: Ib8921224d54cd638b8caab53859c6840791dfd56