Handle custom properties in color attributes
In response to the requirements outlined in Phabricator task T360562,
this commit introduces the CustomPropertyMatcher class aimed at
enhancing the handling and validation of CSS custom properties (variables).
By aligning with the W3C specification for CSS variables,
this addition seeks to bolster the CSS-Sanitizer's capabilities,
ensuring robust and standards-compliant processing of CSS files.

Key Changes:
 - Added CustomPropertyMatcher.php within the Wikimedia\CSS\Grammar
namespace. This class employs regex to validate CSS custom properties,
targeting variable identifiers that adhere to the '--' prefix convention.
This method of validation is crucial for the accurate parsing and interpretation of CSS variables,
ensuring compliance with the W3C CSS Variables Specification.
 - Updated MatcherFactory.php by integrating the new
CustomPropertyMatcher with the existing FunctionMatcher for the 'var'
function in `colorFuncs()`.
- Test cases for the allowed and not allowed cases for
var( --css-prop-name ) in colors

Rationale:
 The development of the CustomPropertyMatcher and its incorporation
into the MatcherFactory.php file directly addresses the needs
identified in Phabricator task T360562. This task highlighted the
necessity for improved support and validation of CSS variables within
the CSS-Sanitizer library, underscoring the importance of this
feature in modern web development practices.

Security:
The attr() and var() functions in CSS are not supported by
css-sanitizer for security purposes.  This patch adds a limited
exception, supporting var() in color attributes.  However,
new values for custom properties can not be declared, limiting
the attack surface.

References:
  - W3C CSS Variables Specification: https://www.w3.org/TR/css-variables-1/

Bug: T360562
Change-Id: I7516a33b415f2327dcca8771fdc33e5968c406e8