docs(css-sanitizer): Update CSS spec references to fixed version
This update modifies the reference links in the documentation comments
to point to a specific version of the CSS variables specification. This
change ensures that the links will remain relevant and accurate over time,
reflecting the exact version of the spec that the current code is based on.

The update follows best practices for documentation by linking to a
specific snapshot of the CSS spec, ensuring that future changes to the
spec do not obscure the context in which the current code was written.

This enhancement was pointed out by Gergő Tisza in a comment on
( https://gerrit.wikimedia.org/r/c/css-sanitizer/+/101465)

Changes:
- `CustomPropertyMatcher.php` and `MatcherFactory.php`: Updated the @see
  links from the mutable URL
  (https://www.w3.org/TR/css-variables-1/#custom-property)
  to the fixed 2022 version
  (https://www.w3.org/TR/2022/CR-css-variables-1-20220616/#custom-property).
  This provides a stable reference for custom CSS properties implementation.

Change-Id: I743d3a434b9f8566e2e721fe9cc7a98ca222d3b3